requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.18.0
streamlit>=1.29.0
praw>=7.7.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
joblib>=1.3.0
anthropic>=0.40.0
feedparser>=6.0.0
yfinance>=0.2.31
bcrypt>=4.1.0
